fix(chat): minor corrections from review
- restore the trailing whitespace on an unrelated breakout comment in
  settings.yml, which this branch had no business touching
- document the mentions settings inline, as the PR description does
- move the vanilla helpers out of the composer and the alert into
  chat-graphql/service.ts: mentionOccurrences, syncPickedMentions,
  pickedMentionsFromHtml, the boundary checks, remapDismissedIndexes and
  parseMessageMetadata
